requirements.txt
Django==1.4.10
PyYAML==3.10
Whoosh==2.5.5
asciitomathml==.89
dj-database-url==0.2.2
dj-static==0.0.5
django-haystack==2.1.0
django-toolbelt==0.0.1
gunicorn==18.0
nltk==2.0.4
psycopg2==2.5.2
pysolr==3.1.0
pystache==0.5.3
requests==2.0.0
scipy==0.13.0
static==1.0.2
wsgiref==0.1.2
